Make `Year` column required with default value in migration
Updated the `ManagementReturns.Year` column to be non-nullable
and set a default value of an empty string. Added a new
migration class `SomePendingThings` to handle this change.
Updated `ReturnsDbContextModelSnapshot` to reflect the new
`Year` property as required with a column type of `nvarchar(max)`.
Included `Up` and `Down` methods in the migration to ensure
reversible changes. Applied namespace and metadata updates
in the migration file.